fix datatype
git-svn-id: https://svn.java.net/svn/grizzly~svn/tags/1_9_18/code@3607 04fdb32c-e91d-0410-84f5-98e489d73e21

+++ modules/config/src/main/java/com/sun/grizzly/config/dom/FileCache.java
@@ -68,7 +68,7 @@ public interface FileCache extends ConfigBeanProxy, Injectable, PropertyBag {
     /**
      * Maximum cache size on the disk
      */
-    @Attribute(defaultValue = "10485760", dataType = Long.class)
+    @Attribute(defaultValue = "10485760", dataType = Integer.class)
     String getMaxCacheSizeBytes();
 
     void setMaxCacheSizeBytes(final String maxCacheSize);
